requirements.txt
anndata==0.10.8
cytocraft==0.0.8
matplotlib==3.7.2
numpy==1.23.5
open3d==0.18.0
pandas==2.2.2
Pillow==10.4.0
plotly==5.22.0
psutil==5.9.8
scipy==1.14.0
seaborn==0.13.2
Shapely==2.0.5
